task1-Tripletex/pyproject.toml
[project]
name = "task1-tripletex"
version = "0.1.0"
description = "Tripletex AI Accounting Agent for NM i AI 2026"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "python-dotenv",
    "openai",
    "openai-agents",
    "openpyxl",
    "pdfplumber>=0.11.9",
]
